moved user values model to different file
