+++ openpype/tools/ayon_push_to_project/models/__init__.py
@@ -1,6 +1,8 @@
 from .selection import PushToProjectSelectionModel
+from .user_values import UserPublishValuesModel
 
 
 __all__ = (
     "PushToProjectSelectionModel",
+    "UserPublishValuesModel",
 )
